Continual restructuring: test suite.
